requirements.txt
streamlit
python-dotenv
PyPDF2
pytesseract
pdf2image
pandas
google-generativeai